refactor endpoints and hooks to separate files, fix abort event, remove console.logs, etc
